Multi device halo ccl - neighborpad (#29212)
### Problem description
In the VAE's for Mochi and Wan 2.2, we had two specific cases where we
needed to be able to exchange boundary values across devices:
1) In order to shard a conv3d op across devices, we needed to shard the
input across H and W, and prepad the input with the correct values from
neighboring devices.
2) When we shard across the T dimension (also in the context of a conv3d
op), we need to be able to pad the front of the time dimension with the
back of the shard from the backward device.

### What's changed
To support this, we created a generic CCL that allows exchanging
boundary values, like a halo op on a single device. For directions that
don't have neighbor devices, zero and replication padding is supported.
This exchange supports different cluster axes, multilink, and exchanging
any dimension.

A barrier semaphore is required instead of using persistent buffers.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/18088337778)
- [x] [T3K
nightly](https://github.com/tenstorrent/tt-metal/actions/runs/18088347003)