#27841 Fix repeat batches (#29887)
### Ticket
#27841 (this should also effect every single model using repeat batches)
### Problem description
- The cause of hang is caused by failing to reset the inputs in between
repeat batches. This is due to the logic introduced in the generator
around `prev_page_table` that controls when we copy the tensors from
host memory to device memory. Namely, the page table, the tokens, the
current position and the rope index.

### What's changed
- A straightforward fix is to explicitly reset the `prev_page_table`
after each repeat batch in our `simple_text_demo.py`.

### Checklist