[skip ci] ci: provide url for every post-commit run (#29944)
### Ticket
None

### Problem description
Right now it's hard to track workflow status manually. When this
workflow gets triggered from tt-llk repo, we get run IDs but not URLs,
so checking on them requires going to tt-metal actions and finding
triggered workflows by hand. This should automate that and make it
easier for the developer to track progress.

This is what we have so far:
<img width="424" height="116" alt="image"
src="https://github.com/user-attachments/assets/6df6d1ef-b506-4497-9175-810611600de7"
/>

With the update, instead of run ID, a user would be served with a URL
leading them to the workflow in question.

### What's changed
Enhanced output of workflow triggering step with URL information.