Support sharding on sequence length for all to all combine for all to all dispatch (#29915)
### Ticket
#29736

### Problem description
Combine currently shards on the batch dimension at the output. For
prefill mode we run with batch=1 and shard on the sequence length, and
thus want the output to shard on sequence length. Similarly, dispatch
currently always concats on the batch dim. Add the ability to
concatenate all the different dispatch device outputs to the sequence
length dimension.

### What's changed
Add the `output_shard_dim` option to `all_to_all_combine` to indicate
which dimension to shard on at the output of combine.
Add the `output_concat_dim` option to `all_to_all_dispatch` to indicate
which dimension to concatenate on at the output of dispatch

### Checklist
- [ ] T3K frequent:
https://github.com/tenstorrent/tt-metal/actions/runs/18204272522