Add back ttnn.reduce_scatter without global semaphores, persistent buffers and automatic decision making for hyperparameters, topology and links (#29985)
### Ticket
#26952

### Problem description
The current reduce scatter is in ttnn.experimental, exposes OP internals
like semaphores and persistent buffers, and makes us pass in hardware
specifications that the user should not have to deal with. As well,
simply removing the global semaphores from the experimental version will
cause failures in existing models as it will increase L1 utilization.

### What's changed
- Add back `ttnn.reduce_scatter` with internalized global semaphore
creation
- Use the same back end as `reduce_scatter_minimal_async` to ensure
optimizations and changes are shared across both
- Keep `reduce_scatter_minimal_async` to ensure people can still tune
the internals as needed.

### What's next
- Incrementally migrate over previous usages to the new one.
- Consider exposing some of the hyperparameters later if we can't find
better defaults in other cases

### Checklist
- T3K: https://github.com/tenstorrent/tt-metal/actions/runs/18236060887
- T3K after rebase:
https://github.com/tenstorrent/tt-metal/actions/runs/18285186735
- APC: https://github.com/tenstorrent/tt-metal/actions/runs/18236066942