fix hang and pcc issues with activation reuse on resnet p150 (#30264)
### Ticket
#30065

### Problem description
Conv that was added to the unit tests was hanging.

### What's changed
- Hang was happening due to the bug with calculating how many cores do
non-meaningful work. We pre-calculate how many last cores process input
data that is 'padding' to fit the sharding scheme. There are N last
cores that do only non-meaningful work, and there is 1 or 0 cores that
do one part of the valid input and one part of 'padding'. This is called
the 'partial core'. In the config we calculate number of cores with
non-meaningful work and this number includes this partial core if it
exists. When calculating starting index of the cores with non-meaningful
work I mistakenly thought that the count includes only cores with only
non-meaningful work.

- While debugging the hang I discovered there was a regression in how we
read the runtime argument in activation reader in
https://github.com/tenstorrent/tt-metal/commit/3dfdd13cbbb0e8501da78efbb5a50a0df2e4343a.
- After the hang was resolved, the test had PCC issues. This was due to
weights double buffering feature - if we have activation reuse enabled,
there is no need to double buffer weights as we keep full inner dim.

### Checklist

- [x] APC:
https://github.com/tenstorrent/tt-metal/actions/runs/18401295188
- [x] Nightly L2:
https://github.com/tenstorrent/tt-metal/actions/runs/18374998139
- [x] Device perf:
https://github.com/tenstorrent/tt-metal/actions/runs/18401184508