[GRIDSAMPLE] Adding allclose check for gridsample nightly tests (#30359)
### Problem description
The pcc check for the gridsample tests is not enough to catch
regression. There is a need for more thorough element-wise check which
is all close. This already exists in some unit tests but needs to be
added in all nightly tests as well.

### What's changed
Added all close check to the gridsample nightly batch of tests with the
same absolute and relative tolerances as already existing all close
check in unit tests

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/18491815738)
- [x] [Nightly tt-metal L2 tests
](https://github.com/tenstorrent/tt-metal/actions/runs/18491786287)