Mstojko/test vision cross attention transformer benchmark (#29699)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/28516

### Problem description
We wanted to add a profiling benchmark for vision cross attention for
gemma3.

### What's changed
Added the `.py` test and a `.json` containing the targets for the
benchmark that are used for the asserts in the test.

Here are the tests for N150, N300 and T3K:
- https://github.com/tenstorrent/tt-metal/actions/runs/18412732701
- https://github.com/tenstorrent/tt-metal/actions/runs/18412742923

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes -
https://github.com/tenstorrent/tt-metal/actions/runs/18221257664

---------

Co-authored-by: Marko Radosavljevic <mradosavljevic@tenstorrent.com>