[AVG POOL 2D] Adding compute kernel config as parameter to avg pool ops  (#30372)
### Ticket
[29149](https://github.com/tenstorrent/tt-metal/issues/29149)

### Problem description
Currently fidelity for the avgpool2d and adaptive_avgpool2d is hardcoded
to HiFi4 where this should be a parameter that can be set in the models
just like it is for conv2d op. Some of the models that require lesser
accuracy can be sped up by lowering fidelity.

### What's changed
Added optional parameter to avg pool and adaptive avg pool op where
compute kernel config can be set. If this is left empty default
parameters are set and HiFi4 is used.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/18408632030)
- [x] [Nightly tt-metal L2 tests
](https://github.com/tenstorrent/tt-metal/actions/runs/18470254812)