[skip ci] Build TT-Train using system packages for TT-Metalium & TT-NN. (#30179)
### Ticket
N/A

### Problem description
TT-Train's current workflow is to build separately from TT-Metalium &
TT-NN. This is brittle when not using exported configs (provided in
packages; not provided in the source tree). Additionally, CI does not
test this flow. The result is pain for the authors.

### What's changed
* Tell TT-Train to first look for system packages and use them if
available
* This is in parallel to the existing
reach-to-many-dirs-because-I-know-secrets approach
* The parallel path should be considered deprecated now and we'll clean
it up after everyone is comfortable with system packages
* Apply the same patches to dependencies that the master repo has
* These are for old CMake versions that ship by default in Ubuntu 22.04
and 24.04
* We copy the patches so that a sparse checkout of ONLY TT-Train has
everything it needs to build cleanly, as an independent project.
* Added a build of TT-Train that specifically uses the .deb packages in
PR Gate to guard against future breakage

### NOTE
This is only testing up to the build phase. Runtime (including tests) is
a future task.