Add errors table to report db (#30386)
### Ticket

https://github.com/tenstorrent/tt-metal/issues/28836

This PR is an updated version of PR #30235, which was merged but had to
be reverted due to the test failing on blackhole.

### Problem description

TTNN Visualizer does not know if it is showing a report from an
incomplete run. The records get written to the db from the decorator
when each operation runs in TTNN, but if an operation raises an
exception, there was no handling of the exception in the decorator to
ensure any error information about the operation was saved to the report
db. The usual information that is saved to the report db for each
operation was also missing, due to the lack of exception handling in the
decorator.

### What's changed

I have fixed the test that was in the reverted PR, and I have ensured it
passes on both wormhole and blackhole.

The test was changed so that instead of just guessing what the maximum
amount of L1 memory available is, it actually checks how much memory is
available first, before creating a tensor that exceeds available memory
in the test case.

--

Original description:

I have wrapped the code that calls the decorated function in
try/except/else/finally blocks. Anything that can still be saved in the
case of an error being raised was moved into the finally block, and
anything that could only be saved to the report db if the decorated
function returned successfully was moved into the else block. In the
except block it will save details of the exception that was raised when
calling the operation to a new errors table in the database.

I have made these changes in the Operation class only, not
FastOperation. In the linked issue, it looks like fast runtime mode is
on (error occurs in FastOperation), but I have made this change only in
the Operation class, i.e. when fast runtime mode is disabled.

I have added a unit test to ensure that the error record gets written
when an exception occurs.

Note that this error record is only getting saved to the database when
there is a runtime exception when calling the operation. It doesn't log
things like syntax errors when running pytest. The test is supposed to
simulate an error of the type someone might want to debug with TTNN
Visualizer - not enough memory.

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI with demo tests passes (if applicable)
- [ ] [Model
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml)
CI passes (if applicable)
- [ ] [Device performance
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml)
CI passes (if applicable)
- [ ] (For models and ops writers) [Single-card demo
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml)
CI passes (if applicable) See [recommended dev
flow](https://github.com/tenstorrent/tt-metal/blob/main/models/docs/MODEL_ADD.md#a-recommended-dev-flow-on-github-for-adding-new-models).
- [ ] [Galaxy
quick](https://github.com/tenstorrent/tt-metal/actions/workflows/galaxy-quick.yaml)
CI passes (if applicable)
- [ ] [Galaxy demo tests, for
Llama](https://github.com/tenstorrent/tt-metal/actions/workflows/galaxy-demo-tests.yaml)
CI passes, if applicable, because of current Llama work
- [ ] (For runtime and ops writers) [T3000 unit
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-unit-tests.yaml)
CI passes (if applicable, since this is run on push to main)
- [ ] (For models and ops writers) [T3000 demo
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-demo-tests.yaml)
CI passes (if applicable, since this is required for release)
- [ ] New/Existing tests provide coverage for changes