Increase upper bound on number of windows that sdpa_windowed supports (#30540)
### Ticket
Link to Github Issue

### Problem description
In Customer Success Team (CST)'s testing of Qwen2.5-VL-3B model, some
test images required more than the current maximum number of windows
(`1037`). The upper bound on the number of windows were experimentally
chosen to support 300DPI images when `sdpa_windowed` was created. The
kernel could support larger numbers of windows. However, we still want
to put an upper bound on that as the performance benefits that
`sdpa_windowed` provides decreases as the number of windows increases.

### What's changed
Change the upper bound number of windows to 4096 (to provide a little
buffer room to grow into). Thanks, @ssanjayTT, for testing and
confirming that this code change works for CST's testing flow.

### Checklist
- [x] T3K unit test CI passes
  - https://github.com/tenstorrent/tt-metal/actions/runs/18502265182
- [x] revert
[0079d70](https://github.com/tenstorrent/tt-metal/pull/30540/commits/0079d7054175f29e493cf75291ac3129fa5656a1)
- through
[11ebd2d](https://github.com/tenstorrent/tt-metal/pull/30540/commits/11ebd2d341aa17c2e49d4739ee17c3d116fb49a8)