Change ttnn python packaging to use scikit-build (#30154)
Closes #26144

- .deb packages and .whl packages have different mechanisms for
identifying what files need to be packaged. This is redundant and
bugprone.
- Many ops in ttnn are not defining what kernel sources are needed at
runtime for `jit`

Provide a single source of truth. Now if it's correct it's correct both
both package formats. If it's wrong, it's wrong for both.

- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/18419904924)
CI passes
- [x] New/Existing tests provide coverage for changes
- [x]
[ClangTidy](https://github.com/tenstorrent/tt-metal/actions/runs/18471772796)
CI passes

---------

Co-authored-by: Andrew Fuller <afuller@tenstorrent.com>