Enabling callstack for active ETH in tt-triage (#30674)
Issues: https://github.com/tenstorrent/tt-exalens/issues/463,
https://github.com/tenstorrent/tt-exalens/issues/399,
https://github.com/tenstorrent/tt-exalens/issues/25

Added support for obtaining callstack of active ERISCs on both wormhole
and blackhole.

In `dispatcher_data.py` script added firmware that are expected to be
loaded on `erisc` on wormhole and `erisc0` and `erisc1` on blackhole.
Additionally, we make distinction between `1-ERISC` and `2-ERISC` mode.
We determine in which mode we are by seeing if
`subordinate_active_erisc.elf` file exist.