Add more verbose message for 404 on WebServer (#30401)
### Ticket
N/A

### Problem description
When running the `tt_telemetry` browser GUI, if the path to which
`$TT_METAL_HOME` is set does not point to the `tt_metal` repository, the
`WebServer` would have returned an uninformative 404.

### What's changed
This PR makes the 404 slightly more informative. What could be done as
well is to check in the `WebServer` constructor not just whether
`$TT_METAL_HOME` is set but also whether it's set correctly. I refrained
from doing that for now to keep things simple, as they were.

### Checklist
- [X] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [X] New/Existing tests provide coverage for changes