Move mesh shape shape reduction to submesh method instead (#30849)
### Problem description
Nightly CCL test was failing due to a trace error during all gather.
However the culprit was an invalid test in all to all dispatch which was
being launched in a sub-grid without the appropriate submesh decoration
which left the trace buffer in the right four devices in the loud box in
a bad state. Then in a later test during all gather when we tried to
create a trace test utilizing those devices, it would claim 0B allocated
as the trace buffer was corrupted.

### What's changed
Fixed the decoration so we create a submesh rather than starting a full
mesh with fewer devices.

https://github.com/tenstorrent/tt-metal/actions/runs/18660916572