Remove dead include in kernel source (#30872)
### Ticket
NA

### Problem description
This is a TTNN kernel that is including a kernel header file from the
tests directory.
This source file <-> test file interaction is problematic to begin with.
Furthermore, nothing in the tests/ directory is packaged, so this should
never work in a clean packaged environment.
Lastly, this header file provides nothing to this translation unit. It
is a dead include.

### What's changed
- rm -rf

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [x] New/Existing tests provide coverage for changes