[skip ci] Bundle SFPI with tarball (again) (#30956)
### Ticket
NA

### Problem description
SFPI either doesn't exist, or is the wrong version on the remote
machines in the multihost workflows.

### What's changed
Switch back to bundling the sfpi compiler along with the build tree.
In the multihost workflow, this effectively installs the current version
of SFPI into the remote machines, because of shared NFS drive.

This PR will give us time to think through how to keep system
dependencies homogeneous on all these machines in mutlihost workflows.