#27100 expose model optimizations to vllm  (#30842)
### Ticket
#27100

vLLM generator hardcodes the model optimizations to compile the model
with. This change makes it configurable when calling
`initialize_vllm_model` and is controlled via the `--override_tt_config`
flag in vLLM server. Accepted values are `performance` and `accuracy`
with fallback to `performance` as default in line with existing
behaviour.

### Checklist
APC: https://github.com/tenstorrent/tt-metal/actions/runs/18679418865
vLLM Nightly:
https://github.com/tenstorrent/tt-metal/actions/runs/18679434495