switch to if constexpr in conv compute kernel (#30665)
#30676

### Problem description
conv compute kernel had a lot of ifdefs in the code that make it hard to
follow the structure; using `if constexpr` is cleaner since it follows
the structure and you can collapse irrelevant code blocks

tip: look at the PR from VSCode extension, indentation changes are
recognized there and you just see actual changes.

### What's changed
- Completely removed ifdefs that weren't used for variable definition;
kept the ones used for definitions only in that part; their occurrences
in rest of the code were replaced with if constexpr
- Renamed file to `conv_bmm_tilize.cpp` to avoid confusion with
col_major that refers to the order in which each core process the
blocks, not to shard orientation of the output
- Renamed `last_out` variable to be more descriptive

Next PR will clean up remaining ifdefs - split reader and activation
reuse; Split into another PR because that one will also remove single
reader + activation reuse functionality as it is not used anymore. This
PR is not meant to change any behavior.

### Checklist

- [x] APC:
https://github.com/tenstorrent/tt-metal/actions/runs/18563810673
- [x] Blackhole PC:
https://github.com/tenstorrent/tt-metal/actions/runs/18563805872
- [x] Frequent models:
https://github.com/tenstorrent/tt-metal/actions/runs/18563800781
- [x] Device perf:
https://github.com/tenstorrent/tt-metal/actions/runs/18587872729
- [x] Model perf:
https://github.com/tenstorrent/tt-metal/actions/runs/18587425497
- [x] L2 nightly:
https://github.com/tenstorrent/tt-metal/actions/runs/18563790577