Delete command processor all gather (#30862)
### Ticket
Status: Open.
#27783 

### Problem description
This op is unmaintained, and we support a superset of its features with
better performance in `ttnn.all_gather`. Every time a test with this op
starts to fail, we simply disable it. With the aggregate/disaggregate
features we added, even the vector of Tensor support can be deprecated.

### What's changed

- Delete all CI testing of the OP on the python side
- Most of the C++ testing is for the Moreh features, which are all
generally disabled at the moment. Those tests need to be refactored with
the `aggregate` + `ttnn.all_gather` + `disaggregate` flow pending some
bug fixes + features. I'm leaving them with the command processor all
gather call commented out so when we refactor we can just insert it in.
- Deleted the all gather command processor implementation.
- Deleted a big chunk of legacy all gather tests that were commented out
and not run

### Checklist
- [ ] APC:
https://github.com/tenstorrent/tt-metal/actions/runs/18690285472
- [ ] T3K:
https://github.com/tenstorrent/tt-metal/actions/runs/18692202244
- [ ] Galaxy:
https://github.com/tenstorrent/tt-metal/actions/runs/18664470946/

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>