Data Format Inf Python Extension  (#752)
### Ticket
<!-- Link to Github Issue -->
https://github.com/tenstorrent/tt-llk/issues/751
### Problem description
<!-- Provide context for the problem. -->
Data format inference currently resides on the C++ side of the test
infrastructure, allowing for full format conversion and configuration
behind the scenes. This was initially designed to let users simply
specify an input and output format in L1, with the rest of the inference
handled automatically.
However, as tests grow more complex—especially with the introduction of
Quasar—this abstraction introduces ambiguity that makes debugging and
configuring tests more difficult. To address this, the exact same data
format inference model has now been moved to the Python side.
### What's changed
<!-- Describe the approach used to solve the problem.
Summarize the changes made and its impact. -->
- All format configurations are explicitly defined ahead of time in the
build.h file, which is imported into the test file at compile time. This
gives us full visibility into the format conversions for each test.
- Additionally, this approach is more flexible as we build out the
inference model for Quasar and extend the interface to support multiple
format combinations per architecture.
- The design enabled these improvements with minimal code changes,
making it easy to extend for Quasar architecture and multiple format
combinations per chip.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update