Verify eth firmware version (#30626)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/30629

### Problem description
For 2 ERISC mode, we decided to not enable it by default to reduce the
impact to customers because the latest Eth Fw is needed for stability.

There are three possible outcomes if you enable try to enable this
feature on an older fw version,

- Things will work
- Initialization will fail. In this case, we already have a message to
tell the user they need to update their tt-firmware version to at least
x.x.x.x
- Something will hang inside the Op. This is the worst case and it's not
easy to debug.

The prevent these undefined behaviors from happening it's important that
we can query the firmware version from Metal and enable/disable/print an
error message as required for this feature.

### What's changed
- Add a check for the eth firmware version

### Checklist
BH
https://github.com/tenstorrent/tt-metal/actions/runs/18724440213
BH Nightly
https://github.com/tenstorrent/tt-metal/actions/runs/18724446395
APC
https://github.com/tenstorrent/tt-metal/actions/runs/18724443005