Fix support for Qwen2.5-VL-7B (#30786)
### Ticket
Link to Github Issue

### Problem description
There is a config error that we have been put off in favor of
higher-priority model variants. Now we are ready to tackle 7B variant.
After investigation, it is due to a unique config of 7B -- it has more
fewer decoder layers in the text part of the model than the vision part
of the model (28 versus 32). In our code, we have been using text
model's `n_layers` to construction optimization strategies for each
decoder layer in both vision and text parts of the model. This bug is
triggered by 7B because `n_layers` of text model is smaller than vision
model's.

### What's changed
This PR adds code that explicitly construct the optimization strategy
for the vision model using the vision model's `n_layers`. This solves
the bug.

### Checklist
- [x] [Single-card demo
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/single-card-demo-tests.yaml)
CI passes
  - https://github.com/tenstorrent/tt-metal/actions/runs/18660962912
- [x] revert
[359be85](https://github.com/tenstorrent/tt-metal/pull/30786/commits/359be8553df9c04785f6f45f027f2a0a52e8608d)
- through
[9777b81](https://github.com/tenstorrent/tt-metal/pull/30786/commits/9777b81d509d426975b8add124fd379f02884a49)
- [x] revert
[36ee689](https://github.com/tenstorrent/tt-metal/pull/30786/commits/36ee6899041da6a4de33a24c135daad7a878323e)
  - through 0a18e7bc57aa55ee65722190aca51054d1133755 
- [x] vLLM nightly
  - https://github.com/tenstorrent/tt-metal/actions/runs/18687081948
- [x] revert 2ce924a557ec53ec53d96cbc195a3c8edf5c89dc
- through
[010cadb](https://github.com/tenstorrent/tt-metal/pull/30786/commits/010cadb60f6971459e92333cbef29d62c5f3c427)