finish generalizing fabric impl to support arbitrarily configured sender channel types (#31130)
Add pathways on host and device to allow both static and elastic
channels to work in the fabric router (either at the same invocation or
different invocations).

This is the last major refactor change needed to the fabric builder and
router code that generalizes the static sized channel datastructure
usage and instantiation. Some of the remaining data-structures affected
by the generalizing to support multiple channel types include:
- sender channel (buffer management/iteration)
- sender channel to producer worker interface
- sender channel to receiver channel flow control managers
- router-to-router adapter 

All of the above are implemented in a way that they can instantiate
either static or elastic channel variants based on compile time args
(i.e. with conditional_t and CRTP).

A substantial new addition with this PR is the additional of the
"MemoryPool". This is defined on host in the builder and is used to
manage router channel allocations. Separate allocators are made
available for elastic and static channels; and the pool can hold one or
more allocators.

To facilitate its use, the memory pool can emit ct args in a way that
can easily be consumed in the kernel.

In kernel code, to collect the CT args from a memory pool you can call:
```c++
using channel_pools_args =
    ChannelPoolCollection<
        ARG_IDX, 
        NUM_SENDER_CHANNELS_IN_POOL, 
        NUM_RECEIVER_CHANNELS_IN_POOL>
```

This will populate some compile datastructures that can be used to
directly build some of the datastructures listed at the start of this
description. For example,

```c++
    auto receiver_channels = tt::tt_fabric::MultiPoolEthChannelBuffers<
        PACKET_HEADER_TYPE,
        channel_pools_args,
        RECEIVER_TO_POOL_TYPE,
        RECEIVER_TO_POOL_IDX>::make();
```

This will construct the tuple of channels (the tuple and be any
combination of static and elastic channel types).

Because the compile time struct builder takes a compile time arg index
as a template argument, it is possible to construct and pass multiple
elastic channel pools. This PR already does this for representing the
local core's pools and remote core pools.

Description Pending

### Sanity checks:
- [x] Galaxy quick:
https://github.com/tenstorrent/tt-metal/actions/runs/18792777005
- [x] T3K Fast:
https://github.com/tenstorrent/tt-metal/actions/runs/18797643811

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI: https://github.com/tenstorrent/tt-metal/actions/runs/18843259147
(T3K Fast after updating to improved perf targets
https://github.com/tenstorrent/tt-metal/actions/runs/18857823917)
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI with demo tests passes (if applicable):
https://github.com/tenstorrent/tt-metal/actions/runs/18843273379
- [x] BH Nightly:
https://github.com/tenstorrent/tt-metal/actions/runs/18795798581
- [x] T3K Nightly:
https://github.com/tenstorrent/tt-metal/actions/runs/18797670333
- [x] Galaxy Nightly:
https://github.com/tenstorrent/tt-metal/actions/runs/18806412907 (old
https://github.com/tenstorrent/tt-metal/actions/runs/18795796336)
- [x] Metal Microbenchmark:
https://github.com/tenstorrent/tt-metal/actions/runs/18795832989