[skip ci] Fix the ccache command when PCH is enabled (#31287)
### Ticket
N/A

### Problem description
The ccache command was malformed when PCH was enabled. The intent was:
> VAR=value ccache ...

However COMPILER_LAUNCHER is not a shell invocation. That doesn't work.
The first argument is expected to be an executable.

Fortunately ccache accepts
> ccache VAR=value ...

### What's changed
Use the format that keeps the executable as the first argument but still
allows us to force overrides on the CLI regardless of the config on the
system.