Revert "nb_util: always place tensor on device when converting in from_numpy" (#31462)
### Ticket
N/A

### Problem description
The reverted change was made in haste to fix what was ultimately a
broken build.

### What's changed
Reverted the original fix attempt, then refactored the code to avoid
double checking the same condition.
Fixed the build with the change reverted and confirmed that the tensor
is making it to the device in from_numpy.

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes