[skip ci] #28487: remove todos after investigation completed (#31416)
### Ticket
Link to Github Issue #28487

### Problem description
A fix to reconfig to make layernorm work correctly caused PCC to be
slightly lower for SDXL. An investigation found that this is not related
to the layernorm op. Given that the PCC changes are small and seeds have
a larger impact, it should be ok to adjust the PCCs.

### What's changed
Remove the TODOs and leave the slightly lower PCCs.