feat: introduce ml_dtypes and switch to it when packing/unpacking data (#784)
### Ticket
None

### Problem description
Doing format conversion requires bit banging and juggling. Using
off-the-shelf lib that does that for us seems like a step in the right
direction. Thus far, we've been doing everything manually. With the MX
formats that will require more juggling. But ml_dtypes already supports
some of them, so that should in the future be easy to implement.
Check https://github.com/jax-ml/ml_dtypes for more.

### What's changed
Switched data packing/unpacking to numpy + ml_dtypes.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update