Added RoPE delta based dynamic mROPE calculation during decode in Qwen2.5-VL models (#31400)
### Ticket
Related to #30980 

### Problem description
We are facing OOMs on Qwen2.5-VL-72B-Instruct because for each user in
the batch, we are allocating cos/sin matrices across the max-model
length. This is giving us a tensor of size `(B, 1, max_model_len,
head_dim)` which occupies too much DRAM and causes OOMs. It also
decreases DRAM throughput during online inferencing because swapping
these large cos and sin tensors from host to device is very slow.

### What's changed
Instead of caching the actual sin/cos values for the multimodal RoPE, we
instead cache only the RoPE deltas. For text tokens generated during
decode, the effective position embedding can be calculated as
`actual_pos_idx=pos_idx+rope_delta` and we can simply index into the cos
and sin matrices as we do in regular RoPE. So we only store one copy of
cos & sin matrices for the entire batch and the rope_deltas which is
essentially a list of B integers. Another advantage is that during
prefill we must not pad the entire sequence of cos and sin matrices to
the entire max_seq_len. Instead we pad it to the max prefill length.

Acompanied vLLM PR: https://github.com/tenstorrent/vllm/pull/228

### Checklist
- [x] (For models and ops writers) [T3000 demo
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/t3000-demo-tests.yaml)
CI passes (if applicable, since this is required for release)
  - https://github.com/tenstorrent/tt-metal/actions/runs/18944205923
- [ ] vLLM nightly
  - https://github.com/tenstorrent/tt-metal/actions/runs/18962415371